backend-ai/requirements.txt
fastapi==0.115.0
uvicorn==0.27.0
python-dotenv==1.0.0
anthropic==0.18.1
openai==1.12.0
python-multipart==0.0.9
httpx==0.26.0
pydantic==2.7.0
requests==2.31.0